Remove unnecessary builder for OpenTelemetry instance
