+++ api/all/src/main/java/io/opentelemetry/api/OpenTelemetry.java
@@ -204,9 +204,4 @@ public interface OpenTelemetry {
 
   /** Returns the {@link ContextPropagators} for this {@link OpenTelemetry}. */
   ContextPropagators getPropagators();
-
-  /** Returns a new {@link OpenTelemetryBuilder}. */
-  static OpenTelemetryBuilder<?> builder() {
-    return new DefaultOpenTelemetry.Builder();
-  }
 }
