+++ api/all/src/main/java/io/opentelemetry/api/Utils.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.api;
+
+import java.util.ServiceLoader;
+import javax.annotation.Nullable;
+
+final class Utils {
+  private Utils() {}
+
+  /**
+   * Load provider class via {@link ServiceLoader}. A specific provider class can be requested via
+   * setting a system property with FQCN.
+   *
+   * @param providerClass a provider class
+   * @param <T> provider type
+   * @return a provider or null if not found
+   * @throws IllegalStateException if a specified provider is not found
+   */
+  @Nullable
+  static <T> T loadSpi(Class<T> providerClass) {
+    String specifiedProvider = System.getProperty(providerClass.getName());
+    ServiceLoader<T> providers = ServiceLoader.load(providerClass);
+    for (T provider : providers) {
+      if (specifiedProvider == null || specifiedProvider.equals(provider.getClass().getName())) {
+        return provider;
+      }
+    }
+    if (specifiedProvider != null) {
+      throw new IllegalStateException(
+          String.format("Service provider %s not found", specifiedProvider));
+    }
+    return null;
+  }
+}
